+++ packages/core-js-compat/src/modules-by-versions.js
@@ -7,5 +7,6 @@ module.exports = {
   ],
   3.2: [
     'es.promise.all-settled',
+    'esnext.symbol.async-dispose',
   ],
 };