+++ packages/core-js/stage/2.js
@@ -1,5 +1,6 @@
 require('../proposals/set-methods');
 require('../proposals/string-replace-all');
 require('../proposals/promise-any');
+require('../proposals/using-statement');
 
 module.exports = require('./3');