+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ##### Unreleased
 - `Promise.allSettled` moved to stable ES, per July TC39 meeting
 - `Promise.any` moved to stage 2, `.errors` property of `AggregateError` instances maked non-enumerable, per July TC39 meeting
+- `using` statement proposal moved to stage 2, added `Symbol.asyncDispose`, per July TC39 meeting
 - Added a fix for [`Math.hypot` V8 7.7 bug](https://bugs.chromium.org/p/v8/issues/detail?id=9546), since it's still not stable without adding results to `core-js-compat`
 - Fixed incorrect early breaking of `{ Map, Set, WeakMap, WeakSet }.deleteAll`
 - Fixed some missed dependencies in entry points